Create Session namespace
Remove Outcomes methods and imports from OneSignalPlugin class, update method names, update iOS header and implementation files
