+++ src/ios/OneSignalPush.h
@@ -97,9 +97,9 @@
 - (void)pauseInAppMessages:(CDVInvokedUrlCommand* _Nonnull)command;
 
 // Outcomes
-- (void)sendOutcome:(CDVInvokedUrlCommand* _Nonnull)command;
-- (void)sendUniqueOutcome:(CDVInvokedUrlCommand* _Nonnull)command;
-- (void)sendOutcomeWithValue:(CDVInvokedUrlCommand* _Nonnull)command;
+- (void)addOutcome:(CDVInvokedUrlCommand* _Nonnull)command;
+- (void)addUniqueOutcome:(CDVInvokedUrlCommand* _Nonnull)command;
+- (void)addOutcomeWithValue:(CDVInvokedUrlCommand* _Nonnull)command;
 
 // Location
 - (void)promptLocation:(CDVInvokedUrlCommand* _Nonnull)command;
