+++ src/ios/OneSignalPush.m
@@ -594,29 +594,23 @@ static Class delegateClass = nil;
  * Outcomes
  */
 
-- (void)sendOutcome:(CDVInvokedUrlCommand*)command {
+- (void)addOutcome:(CDVInvokedUrlCommand*)command {
     NSString *name = command.arguments[0];
 
-    [OneSignal sendOutcome:name onSuccess:^(OSOutcomeEvent *outcome){
-        successCallback(command.callbackId, [outcome jsonRepresentation]);
-    }];
+    [OneSignal.Session addOutcome:name];
 }
 
-- (void)sendUniqueOutcome:(CDVInvokedUrlCommand*)command {
+- (void)addUniqueOutcome:(CDVInvokedUrlCommand*)command {
     NSString *name = command.arguments[0];
 
-    [OneSignal sendUniqueOutcome:name onSuccess:^(OSOutcomeEvent *outcome){
-        successCallback(command.callbackId, [outcome jsonRepresentation]);
-    }];
+    [OneSignal.Session addUniqueOutcome:name];
 }
 
-- (void)sendOutcomeWithValue:(CDVInvokedUrlCommand*)command {
+- (void)addOutcomeWithValue:(CDVInvokedUrlCommand*)command {
     NSString *name = command.arguments[0];
     NSNumber *value = command.arguments[1];
-
-    [OneSignal sendOutcomeWithValue:name value:value onSuccess:^(OSOutcomeEvent *outcome){
-        successCallback(command.callbackId, [outcome jsonRepresentation]);
-    }];
+    
+    [OneSignal.Session addOutcomeWithValue:name value:value];
 }
 
 /**
